restructuring chaindb.go and groupdb.go blockdb.go
