+++ internal/pkg/chainsdk/core/molassessnapshotreceiver.go
@@ -138,7 +138,7 @@ func (ssreceiver *MolassesSnapshotReceiver) doApply(snapshots map[string]*quorum
 					return err
 				}
 			} else if snapshotdata.Type == quorumpb.SnapShotItemType_SNAPSHOT_CHAIN_CONFIG {
-				err := nodectx.GetDbMgr().UpdateChainConfig(snapshotdata.Data, ssreceiver.nodename)
+				err := nodectx.GetNodeCtx().GetChainStorage().UpdateChainConfig(snapshotdata.Data, ssreceiver.nodename)
 				if err != nil {
 					snapshotreceiver_log.Warningf("<%s> applySnapshot failed, type CHAIN_CONFIG, err <%s>", ssreceiver.groupId, err.Error())
 					return err
