+++ internal/pkg/storage/def/chainstorage.go
@@ -1,0 +1,11 @@
+package def
+
+import (
+	quorumpb "github.com/rumsystem/rumchaindata/pkg/pb"
+)
+
+type ChainStorageIface interface {
+	DeleteRelay(relayid string) (bool, *quorumpb.GroupRelayItem, error)
+	AddRelayActivity(groupRelayItem *quorumpb.GroupRelayItem) (string, error)
+	AddRelayReq(groupRelayItem *quorumpb.GroupRelayItem) (string, error)
+}
